+++ packages/playground/glob-import/__tests__/glob-import.spec.ts
@@ -45,8 +45,14 @@ const allResult = {
   }
 }
 
-const resultNodeModules = {
+const nodeModulesResult = {
   '/dir/node_modules/hoge.js': { msg: 'hoge' }
+}
+
+const rawResult = {
+  '/dir/baz.json': {
+    msg: 'baz'
+  }
 }
 
 test('should work', async () => {
@@ -54,7 +60,13 @@ test('should work', async () => {
     JSON.stringify(allResult, null, 2)
   )
   expect(await page.textContent('.result-node_modules')).toBe(
-    JSON.stringify(resultNodeModules, null, 2)
+    JSON.stringify(nodeModulesResult, null, 2)
+  )
+})
+
+test('import glob raw', async () => {
+  expect(await page.textContent('.globraw')).toBe(
+    JSON.stringify(rawResult, null, 2)
   )
 })