+++ packages/playground/vue/Main.vue
@@ -18,8 +18,9 @@
   <Suspense>
     <AsyncComponent />
   </Suspense>
-  <RefTransform />
+  <ReactivityTransform :foo="time" />
   <SetupImportTemplate />
+  <ExternalStyleCss />
 </template>
 
 <script setup lang="ts">
@@ -33,8 +34,9 @@ import SrcImport from './src-import/SrcImport.vue'
 import Slotted from './Slotted.vue'
 import ScanDep from './ScanDep.vue'
 import AsyncComponent from './AsyncComponent.vue'
-import RefTransform from './RefTransform.vue'
+import ReactivityTransform from './ReactivityTransform.vue'
 import SetupImportTemplate from './setup-import-template/SetupImportTemplate.vue'
+import ExternalStyleCss from './ExternalStyleCss.vue'
 
 import { ref } from 'vue'