+++ packages/playground/worker/__tests__/worker.spec.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 import { untilUpdated, isBuild, testDir } from '../../testUtils'
-import { Page } from 'playwright-chromium'
+import type { Page } from 'playwright-chromium'
 
 test('normal', async () => {
   await page.click('.ping')
@@ -9,6 +9,10 @@ test('normal', async () => {
   await untilUpdated(
     () => page.textContent('.mode'),
     isBuild ? 'production' : 'development'
+  )
+  await untilUpdated(
+    () => page.textContent('.bundle-with-plugin'),
+    'worker bundle with plugin success!'
   )
 })
 
@@ -48,12 +52,11 @@ test.concurrent.each([[true], [false]])('shared worker', async (doTick) => {
 })
 
 if (isBuild) {
+  const assetsDir = path.resolve(testDir, 'dist/assets')
   // assert correct files
   test('inlined code generation', async () => {
-    const assetsDir = path.resolve(testDir, 'dist/assets')
     const files = fs.readdirSync(assetsDir)
-    // should have 3 worker chunk
-    expect(files.length).toBe(4)
+    expect(files.length).toBe(6)
     const index = files.find((f) => f.includes('index'))
     const content = fs.readFileSync(path.resolve(assetsDir, index), 'utf-8')
     const worker = files.find((f) => f.includes('my-worker'))
@@ -71,5 +74,17 @@ if (isBuild) {
     // inlined
     expect(content).toMatch(`(window.URL||window.webkitURL).createObjectURL`)
     expect(content).toMatch(`window.Blob`)
+  })
+
+  test('worker need bundle', () => {
+    fs.readdirSync(assetsDir)
+      .filter(
+        (file) =>
+          file.includes('url-worker') || file.includes('url-shared-worker')
+      )
+      .forEach((file) => {
+        const content = fs.readFileSync(path.resolve(assetsDir, file), 'utf-8')
+        expect(content.startsWith('(function(){')).toBe(true)
+      })
   })
 }