+++ packages/vite/src/node/__tests__/config.spec.ts
@@ -1,11 +1,6 @@
-import { InlineConfig } from '..'
-import {
-  mergeConfig,
-  resolveConfig,
-  UserConfigExport,
-  resolveEnvPrefix,
-  UserConfig
-} from '../config'
+import type { InlineConfig } from '..'
+import type { UserConfigExport, UserConfig } from '../config'
+import { mergeConfig, resolveConfig, resolveEnvPrefix } from '../config'
 
 describe('mergeConfig', () => {
   test('handles configs with different alias schemas', () => {
@@ -46,6 +41,22 @@ describe('mergeConfig', () => {
           }
         ]
       }
+    }
+
+    expect(mergeConfig(baseConfig, newConfig)).toEqual(mergedConfig)
+  })
+
+  test('handles arrays', () => {
+    const baseConfig: UserConfigExport = {
+      envPrefix: 'string1'
+    }
+
+    const newConfig: UserConfigExport = {
+      envPrefix: ['string2', 'string3']
+    }
+
+    const mergedConfig: UserConfigExport = {
+      envPrefix: ['string1', 'string2', 'string3']
     }
 
     expect(mergeConfig(baseConfig, newConfig)).toEqual(mergedConfig)
@@ -172,6 +183,9 @@ describe('preview config', () => {
     host: true,
     open: true,
     https: true,
+    headers: {
+      'Cache-Control': 'no-store'
+    },
     proxy: { '/foo': 'http://localhost:4567' },
     cors: false
   })