+++ packages/vite/src/node/importGlob.ts
@@ -1,12 +1,20 @@
 import path from 'path'
+import { promises as fsp } from 'fs'
 import glob from 'fast-glob'
+import * as JSON5 from 'json5'
 import {
   isModernFlag,
   preloadMethod,
   preloadMarker
 } from './plugins/importAnalysisBuild'
 import { cleanUrl } from './utils'
-import { RollupError } from 'rollup'
+import type { RollupError } from 'rollup'
+
+export interface AssertOptions {
+  assert?: {
+    type: string
+  }
+}
 
 export async function transformImportGlob(
   source: string,
@@ -38,7 +46,7 @@ export async function transformImportGlob(
   importer = cleanUrl(importer)
   const importerBasename = path.basename(importer)
 
-  let [pattern, endIndex] = lexGlobPattern(source, pos)
+  let [pattern, assertion, endIndex] = lexGlobPattern(source, pos)
   if (!pattern.startsWith('.') && !pattern.startsWith('/')) {
     throw err(`pattern must start with "." or "/" (relative to project root)`)
   }
@@ -80,8 +88,12 @@ export async function transformImportGlob(
       ;[importee] = await normalizeUrl(file, pos)
     }
     imports.push(importee)
-    const identifier = `__glob_${importIndex}_${i}`
-    if (isEager) {
+    if (assertion?.assert?.type === 'raw') {
+      entries += ` ${JSON.stringify(file)}: ${JSON.stringify(
+        await fsp.readFile(path.join(base, file), 'utf-8')
+      )},`
+    } else if (isEager) {
+      const identifier = `__glob_${importIndex}_${i}`
       importsString += `import ${
         isEagerDefault ? `` : `* as `
       }${identifier} from ${JSON.stringify(importee)};`
@@ -116,7 +128,10 @@ const enum LexerState {
   inTemplateString
 }
 
-function lexGlobPattern(code: string, pos: number): [string, number] {
+function lexGlobPattern(
+  code: string,
+  pos: number
+): [string, AssertOptions, number] {
   let state = LexerState.inCall
   let pattern = ''
 
@@ -162,7 +177,45 @@ function lexGlobPattern(code: string, pos: number): [string, number] {
         throw new Error('unknown import.meta.glob lexer state')
     }
   }
-  return [pattern, code.indexOf(`)`, i) + 1]
+
+  const endIndex = getEndIndex(code, i)
+  const options = code.substring(i + 1, endIndex)
+  const commaIndex = options.indexOf(`,`)
+  let assert = {}
+  if (commaIndex > -1) {
+    assert = JSON5.parse(options.substr(commaIndex + 1))
+  }
+  return [pattern, assert, endIndex + 1]
+}
+
+// reg without the 'g' option, only matches the first match
+const multilineCommentsRE = /\/\*(.|[\r\n])*?\*\//m
+const singlelineCommentsRE = /\/\/.*/
+
+function getEndIndex(code: string, i: number): number {
+  const findStart = i
+  const endIndex = code.indexOf(`)`, findStart)
+  const subCode = code.substring(findStart)
+
+  const matched =
+    subCode.match(singlelineCommentsRE) ?? subCode.match(multilineCommentsRE)
+  if (!matched) {
+    return endIndex
+  }
+
+  const str = matched[0]
+  const index = matched.index
+  if (!index) {
+    return endIndex
+  }
+
+  const commentStart = findStart + index
+  const commentEnd = commentStart + str.length
+  if (endIndex > commentStart && endIndex < commentEnd) {
+    return getEndIndex(code, commentEnd)
+  } else {
+    return endIndex
+  }
 }
 
 function error(pos: number) {