+++ packages/vite/src/node/server/openBrowser.ts
@@ -10,10 +10,10 @@
 
 import path from 'path'
 import open from 'open'
-import execa from 'execa'
-import chalk from 'chalk'
+import spawn from 'cross-spawn'
+import colors from 'picocolors'
 import { execSync } from 'child_process'
-import { Logger } from '../logger'
+import type { Logger } from '../logger'
 
 // https://github.com/sindresorhus/open#app
 const OSX_CHROME = 'google chrome'
@@ -40,14 +40,14 @@ export function openBrowser(
 
 function executeNodeScript(scriptPath: string, url: string, logger: Logger) {
   const extraArgs = process.argv.slice(2)
-  const child = execa('node', [scriptPath, ...extraArgs, url], {
+  const child = spawn(process.execPath, [scriptPath, ...extraArgs, url], {
     stdio: 'inherit'
   })
   child.on('close', (code) => {
     if (code !== 0) {
       logger.error(
-        chalk.red(
-          `\nThe script specified as BROWSER environment variable failed.\n\n${chalk.cyan(
+        colors.red(
+          `\nThe script specified as BROWSER environment variable failed.\n\n${colors.cyan(
             scriptPath
           )} exited with code ${code}.`
         ),